Fix fallback when `undefined` is explicitly implemented
Previously transmute had a bug where the fallback implementation would never be used if an implementation for `undefined` was implemented.
This is because when protocol.js's dispatch function calls getValueKey, it will correctly get `undefined`, which is correct for an unimplemented type of value.
Unfortunately, implementations[key] on the next line will coerce that `undefined` into the string `"undefined"` to look up the implementation, where it
will find the implementation for when the value is exactly `undefined`. This is why for all cases where the fallback implementation would have been used,
`dispatch` was using the implementation for the exact value `undefined`.

To solve this, we check for the existence of the key first before attempting to look it up in the implementations object.